Refactored utils.ts becouse of error in api legacy linking dist
