+++ src/rankify/__tests__/InstanceBase.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, jest } from "@jest/globals";
 import { type PublicClient, type GetContractEventsReturnType, type Hex } from "viem";
 import InstanceBase from "../InstanceBase";
-import { MOCK_ADDRESSES, MOCK_HASHES, createMockPublicClient } from "../../__tests__/utils";
+import { MOCK_ADDRESSES, MOCK_HASHES, createMockPublicClient } from "../../__tests__/utils.test";
 
 // Mock viem
 jest.mock("viem", () => ({
